ref: Move integration out of `opentelemetry/integrations/`
