+++ sentry_sdk/_init_implementation.py
@@ -3,6 +3,10 @@ from typing import TYPE_CHECKING
 import sentry_sdk
 from sentry_sdk.consts import ClientConstructor
 from sentry_sdk.opentelemetry.scope import setup_scope_context_management
+from sentry_sdk.opentelemetry.integration import (
+    patch_readable_span,
+    setup_sentry_tracing,
+)
 
 if TYPE_CHECKING:
     from typing import Any, Optional
@@ -22,6 +26,8 @@ def _init(*args, **kwargs):
 
     This takes the same arguments as the client constructor.
     """
+    patch_readable_span()
+    setup_sentry_tracing()
     setup_scope_context_management()
     client = sentry_sdk.Client(*args, **kwargs)
     sentry_sdk.get_global_scope().set_client(client)
